pyproject.toml
[project]
name = "agdg"
version = "0.1.0"
dependencies = [
	"accelerate",
	"boto3",
	"datasets",
	"diffusers",
	"huggingface_hub",
	"modal",
	"numpy",
	"pillow",
	"psycopg2-binary",
	"pytest",
	"pytest-cov",
	"python-dotenv",
	"sentence-transformers",
	"tf-keras",
	"torch",
	"tqdm",
	"transformers",
]

[tool.setuptools.packages.find]
where = ["src"]
